// src/recipe/dto/filter-recipe.dto.ts
export class GetRecipiesFilterDto {
    name?: string;
    author?: string;
    avg_rating?: number;
    is_public?: any;
    ingredients?: string;
    meal_type?: string;
    country?: string;
    cooking_time?: number;
    difficulty?: string;
    views?: number;
    food_type?: string;
}